@@ -1,6 +1,6 @@
 {
   "name": "polkaswap-exchange-web",
-  "version": "1.20.0",
+  "version": "1.21.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/sora-xor/polkaswap-exchange-web.git"
@@ -13,6 +13,7 @@
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint",
+    "inspect": "vue-cli-service inspect > .webpack-config.js",
     "electron:build": "vue-cli-service electron:build",
     "electron:serve": "vue-cli-service electron:serve",
     "generate-lang-json": "ts-node --skip-project ./scripts/generateLocaleJson.ts",
@@ -23,20 +24,19 @@
   },
   "dependencies": {
     "@metamask/detect-provider": "^2.0.0",
-    "@soramitsu/soraneo-wallet-web": "1.20.0",
+    "@soramitsu/soraneo-wallet-web": "1.21.10",
     "@walletconnect/web3-provider": "^1.8.0",
-    "base-64": "^1.0.0",
     "core-js": "^3.26.0",
     "country-code-emoji": "^2.3.0",
     "country-flag-emoji-polyfill": "^0.1.4",
     "direct-vuex": "^0.12.1",
     "echarts": "^5.4.1",
     "email-validator": "^2.0.4",
-    "ethers": "^5.7.2",
+    "ethers": "6.7.1",
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.21",
     "uuidv4": "^6.2.13",
-    "vue": "2.6.14",
+    "vue": "2.7.14",
     "vue-class-component": "^7.2.6",
     "vue-echarts": "^6.3.3",
     "vue-i18n": "^8.11.2",
@@ -45,23 +45,22 @@
     "vuex": "^3.1.3"
   },
   "devDependencies": {
-    "@babel/runtime": "^7.20.0",
+    "@babel/runtime": "^7.22.11",
     "@types/electron-devtools-installer": "^2.2.2",
     "@types/jest": "^29.2.0",
-    "@types/jsdom": "^20.0.0",
+    "@types/jsdom": "^21.1.1",
     "@types/lodash": "^4.14.186",
-    "@types/node": "^18.11.7",
+    "@types/node": "^20.5.6",
     "@typescript-eslint/eslint-plugin": "^5.2.0",
     "@typescript-eslint/parser": "^5.2.0",
-    "@vue/cli-plugin-babel": "^5.0.0-beta.4",
-    "@vue/cli-plugin-e2e-cypress": "^5.0.0-beta.4",
-    "@vue/cli-plugin-eslint": "^5.0.0-beta.4",
-    "@vue/cli-plugin-router": "^5.0.0-beta.4",
-    "@vue/cli-plugin-typescript": "^5.0.0-beta.4",
-    "@vue/cli-plugin-unit-jest": "^5.0.0-beta.4",
-    "@vue/cli-plugin-vuex": "^5.0.0-beta.4",
-    "@vue/cli-service": "^5.0.0-beta.4",
-    "@vue/composition-api": "^1.7.1",
+    "@vue/cli-plugin-babel": "5.0.8",
+    "@vue/cli-plugin-e2e-cypress": "5.0.8",
+    "@vue/cli-plugin-eslint": "5.0.8",
+    "@vue/cli-plugin-router": "5.0.8",
+    "@vue/cli-plugin-typescript": "5.0.8",
+    "@vue/cli-plugin-unit-jest": "5.0.8",
+    "@vue/cli-plugin-vuex": "5.0.8",
+    "@vue/cli-service": "5.0.8",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-standard": "^6.1.0",
     "@vue/eslint-config-typescript": "^8.0.0",
@@ -83,14 +82,13 @@
     "lint-staged": "^9.5.0",
     "node-polyfill-webpack-plugin": "^2.0.1",
     "prettier": "^2.2.1",
-    "sass": "^1.55.0",
-    "sass-loader": "^13.1.0",
+    "sass": "^1.66.1",
+    "sass-loader": "^13.3.2",
     "ts-jest": "^27.0.5",
-    "typescript": "~4.4.4",
+    "typescript": "~4.9.5",
     "vue-cli-plugin-electron-builder": "^3.0.0-alpha.3",
     "vue-cli-plugin-test-attrs": "^0.1.5",
-    "vue-svg-loader": "^0.16.0",
-    "vue-template-compiler": "2.6.14"
+    "vue-template-compiler": "2.7.14"
   },
   "resolutions": {
     "@polkadot/extension-dapp": "^0.44.7",
