@@ -51,6 +51,12 @@
         <span>{{ statisticsConnectionDesc }}</span>
       </template>
     </footer-popper>
+    <div class="sora-logo">
+      <span class="sora-logo__title">{{ t('poweredBy') }}</span>
+      <a class="sora-logo__image" href="https://sora.org" title="Sora" target="_blank" rel="nofollow noopener">
+        <sora-logo :theme="libraryTheme" />
+      </a>
+    </div>
     <select-node-dialog />
     <no-internet-dialog />
     <statistics-dialog :visible.sync="showStatisticsDialog" />
@@ -64,6 +70,7 @@ import { getExplorerLinks, WALLET_TYPES } from '@soramitsu/soraneo-wallet-web';
 import { Component, Mixins } from 'vue-property-decorator';
 
 import TranslationMixin from '@/components/mixins/TranslationMixin';
+import SoraLogo from '@/components/shared/Logo/Sora.vue';
 import { Components } from '@/consts';
 import { lazyComponent } from '@/router';
 import { state, getter, mutation } from '@/store/decorators';
@@ -73,6 +80,7 @@ import FooterPopper from './FooterPopper.vue';
 import { formatLocation } from './Node/utils';
 import NoInternetDialog from './NoInternetDialog.vue';
 
+import type Theme from '@soramitsu/soramitsu-js-ui/lib/types/Theme';
 import type { WALLET_CONSTS } from '@soramitsu/soraneo-wallet-web';
 
 /** Max limit provided by navigator.connection.downlink */
@@ -80,6 +88,7 @@ const MAX_INTERNET_CONNECTION_LIMIT = 10;
 
 @Component({
   components: {
+    SoraLogo,
     FooterPopper,
     NoInternetDialog,
     StatisticsDialog: lazyComponent(Components.StatisticsDialog),
@@ -90,6 +99,7 @@ export default class AppFooter extends Mixins(TranslationMixin) {
   // Block explorer
   @state.wallet.settings.soraNetwork private soraNetwork!: Nullable<WALLET_CONSTS.SoraNetwork>;
   @state.settings.blockNumber blockNumber!: number;
+  @getter.libraryTheme libraryTheme!: Theme;
 
   get blockExplorerLink(): Nullable<string> {
     const links = getExplorerLinks(this.soraNetwork);
@@ -104,10 +114,14 @@ export default class AppFooter extends Mixins(TranslationMixin) {
   }
 
   // Node connection
-  @getter.settings.nodeIsConnecting isNodeConnecting!: boolean;
+  @state.settings.node connectedNode!: Partial<Node>;
+  @getter.settings.connectingNode connectingNode!: Nullable<Node>;
   @getter.settings.nodeIsConnected isNodeConnected!: boolean;
-  @state.settings.node node!: Partial<Node>;
   @mutation.settings.setSelectNodeDialogVisibility private setSelectNodeDialogVisibility!: (flag: boolean) => void;
+
+  private get isNodeConnecting(): boolean {
+    return !!this.connectingNode;
+  }
 
   private get nodeConnectionStatusKey(): string {
     if (this.isNodeConnected) return 'connected';
@@ -117,12 +131,16 @@ export default class AppFooter extends Mixins(TranslationMixin) {
 
   get nodeConnectionClass(): Status {
     if (this.isNodeConnected) return Status.SUCCESS;
-    if (this.isNodeConnecting) return Status.DEFAULT;
+    if (this.isNodeConnecting) return Status.INFO;
     return Status.ERROR;
   }
 
   get nodeConnectionText(): string {
     return this.t(`footer.node.title.${this.nodeConnectionStatusKey}`);
+  }
+
+  get node() {
+    return this.connectingNode ?? this.connectedNode;
   }
 
   get formattedNodeLocation() {
@@ -184,11 +202,11 @@ export default class AppFooter extends Mixins(TranslationMixin) {
       case WALLET_TYPES.ConnectionStatus.Unavailable:
         return Status.ERROR;
       case WALLET_TYPES.ConnectionStatus.Loading:
-        return Status.DEFAULT;
+        return Status.INFO;
       case WALLET_TYPES.ConnectionStatus.Available:
         return Status.SUCCESS;
       default:
-        return Status.DEFAULT;
+        return Status.INFO;
     }
   }
 
@@ -233,6 +251,8 @@ export default class AppFooter extends Mixins(TranslationMixin) {
 
 <style lang="scss" scoped>
 $block-icon-size: 7px;
+$sora-logo-height: 36px;
+$sora-logo-width: 115px;
 
 .app-status {
   font-size: var(--s-font-size-extra-mini);
@@ -254,6 +274,37 @@ $block-icon-size: 7px;
     height: $block-icon-size;
     width: $block-icon-size;
     margin-right: 2px;
+  }
+}
+
+.sora-logo {
+  display: flex;
+  align-items: center;
+  align-self: flex-end;
+  position: absolute;
+  right: 0;
+
+  &__title {
+    text-transform: uppercase;
+    font-weight: 200;
+    color: var(--s-color-base-content-secondary);
+    font-size: 12px;
+    line-height: 12px;
+    margin-right: $basic-spacing / 2;
+    margin-top: 2px;
+    white-space: nowrap;
+  }
+
+  &__image {
+    width: $sora-logo-width;
+    height: $sora-logo-height;
+    @include focus-outline;
+  }
+}
+
+@include desktop(true) {
+  .sora-logo {
+    display: none;
   }
 }
 </style>
