@@ -1,5 +1,5 @@
 <template>
-  <svg width="174" height="36" viewBox="0 0 174 36" fill="none" xmlns="http://www.w3.org/2000/svg">
+  <svg width="98" height="36" viewBox="0 0 174 36" fill="none" xmlns="http://www.w3.org/2000/svg">
     <path
       d="M103.478 19.8651H105.913V18.7374H103.478V19.8651ZM151.13 24.0289L153.043 23.3349C152.348 21.6 151.565 19.9518 150.696 18.2169L148.87 18.9976C149.739 20.906 150.522 22.5542 151.13 24.0289ZM158.609 18.0434C158.522 19.6916 158.435 21.0795 158.087 22.2072C157.826 23.3349 157.304 24.2024 156.696 24.9831C156.087 25.7639 155.304 26.2843 154.348 26.7181C153.391 27.1518 152.174 27.412 150.696 27.5855L151.043 29.4072C153.304 29.147 155.13 28.5398 156.435 27.759C157.739 26.9783 158.783 25.7639 159.391 24.2024C160 22.641 160.435 20.6458 160.435 18.0434H158.609ZM162.348 23.6819H171.826C171.565 25.0699 170.783 26.1108 169.565 26.8048C168.348 27.4988 166.609 27.8458 164.522 28.0193L164.783 29.7542C167.913 29.5807 170.261 28.8867 171.739 27.6723C173.217 26.4578 174 24.6361 174 22.294V21.947H162.435L162.348 23.6819ZM173.043 19.8651V18.1301H163.391V19.8651H173.043ZM9.47826 33.8313C12 35.2193 14.9565 36 18.087 36C21.2174 36 24.087 35.2193 26.6957 33.8313L18.087 21.0795L9.47826 33.8313ZM7.04348 17.9566V15.0072H16.6087V11.9711H7.04348V9.02169H29.1304V11.9711H19.5652V15.0072H29.1304V17.9566H19.5652L29.1304 32.1831C33.3913 28.8867 36.087 23.7687 36.087 17.9566C36.087 8.06747 28 0 18.087 0C8.08696 0 0 8.06747 0 17.9566C0 23.7687 2.69565 28.8867 6.95652 32.1831L16.5217 17.9566H7.04348ZM54 30.3614C59.5652 30.3614 63.2174 27.412 63.2174 23.0747C63.2174 19.0843 60 16.8289 55.1304 16.3952L52.9565 16.2217C50.6087 16.0482 49.4783 14.747 49.4783 13.2723C49.4783 11.3639 50.8696 9.71566 53.8261 9.71566C56.7826 9.71566 58.1739 11.5373 58.1739 13.4458H62.1739C62.1739 9.02169 58.8696 6.15904 53.8261 6.15904C48.7826 6.15904 45.5652 8.93494 45.5652 13.2723C45.5652 17.0024 48.3478 19.5181 52.6087 19.8651L54.7826 20.0386C57.7391 20.2988 59.2174 21.4265 59.2174 23.0747C59.2174 25.1566 57.5652 26.8048 54 26.8048C50.087 26.8048 48.6957 24.8096 48.6957 22.9012H44.6957C44.6957 27.2386 48.087 30.3614 54 30.3614ZM78.9565 30.3614C87.2174 30.3614 91.0435 23.6819 91.0435 18.5639V17.8699C91.0435 12.3181 87.0435 6.15904 78.9565 6.15904C70.8696 6.15904 66.8696 12.3181 66.8696 17.8699V18.5639C66.8696 23.6819 70.6956 30.3614 78.9565 30.3614ZM78.9565 26.6313C74.0869 26.6313 70.9565 22.7277 70.9565 18.3036C70.9565 13.6193 74 9.9759 78.9565 9.9759C83.8261 9.9759 86.8696 13.6193 86.8696 18.3036C86.8696 22.7277 83.8261 26.6313 78.9565 26.6313ZM110.87 29.7542H115.652L109.391 21.253C112.174 20.212 113.826 18.0434 113.826 14.5735V14.053C113.826 9.02169 110.435 6.67952 105.043 6.67952H96.8696V29.7542H100.957V21.947H105.13H105.391L110.87 29.7542ZM100.783 10.1494H105.478C108.087 10.1494 109.739 11.8843 109.739 14.3133C109.739 16.8289 108.087 18.4771 105.478 18.4771H100.783V10.1494ZM136.174 29.7542H140.348L132.174 6.85301H125.478L117.478 29.7542H121.565L123.739 23.4217H134L136.174 29.7542ZM124.87 20.0386L128.174 10.3229H129.304L132.696 20.0386H124.87Z"
       :fill="textColor"
