@@ -3,11 +3,10 @@ import { RewardingEvents } from '@sora-substrate/util/build/rewards/consts';
 import Theme from '@soramitsu/soramitsu-js-ui/lib/types/Theme';
 import { en as walletEn } from '@soramitsu/soraneo-wallet-web';
 
-import { AlertFrequencyTabs, AlertTypeTabs } from '@/types/tabs';
-
 import { MoonpayNotifications } from '../components/pages/Moonpay/consts';
 import { PageNames, RewardsTabsItems } from '../consts';
 import { DemeterPageNames } from '../modules/demeterFarming/consts';
+import { AlertFrequencyTabs, AlertTypeTabs, FiatOptionTabs } from '../types/tabs';
 
 export default {
   // Wallet project keys
@@ -118,6 +117,17 @@ export default {
     frequencyTooltip:
       "Select between 'once' and 'always' to determine how often you receive notifications for the chosen alert type. 'Once' will send a single notification when the condition is met, while 'always' will continue to notify you each time the price threshold is crossed.",
     noSupportMsg: "Notifications aren't supported by your browser",
+  },
+  fiatPayment: {
+    [FiatOptionTabs.moonpay]: 'MoonPay',
+    [FiatOptionTabs.x1ex]: 'x1ex',
+    historyBtn: 'My purchases',
+    historyTitle: 'Transaction History',
+    moonpayTitle: 'Buy ETH via MoonPay',
+    moonpayDesc: 'Purchase ETH tokens on Ethereum and transfer them to SORA network via the bridge',
+    x1Title: 'Buy XOR with your card',
+    x1Desc: 'Purchase XOR tokens with your debit or credit card',
+    x1Btn: 'Buy XOR with card',
   },
   headerMenu: {
     showBalances: 'Show Balances',
