@@ -21,9 +21,12 @@ const getters = defineGetters<SettingsState>()({
     const { state, getters } = settingsGetterContext(args);
     return [...state.defaultNodes, ...getters.customNodes];
   },
-  nodeIsConnecting(...args): boolean {
-    const { state } = settingsGetterContext(args);
-    return !!state.nodeAddressConnecting;
+  connectingNode(...args): Nullable<Node> {
+    const { state, getters } = settingsGetterContext(args);
+
+    if (!state.nodeAddressConnecting) return null;
+
+    return getters.nodeList.find((node) => node.address === state.nodeAddressConnecting);
   },
   nodeIsConnected(...args): boolean {
     const { state } = settingsGetterContext(args);
@@ -40,6 +43,14 @@ const getters = defineGetters<SettingsState>()({
   moonpayEnabled(...args): boolean {
     const { state, getters } = settingsGetterContext(args);
     return !!getters.moonpayApiKey && !!state.featureFlags.moonpay;
+  },
+  x1ApiKey(...args): string {
+    const { rootState } = settingsGetterContext(args);
+    return rootState.wallet.settings.apiKeys.x1ex;
+  },
+  x1Enabled(...args): boolean {
+    const { state, getters } = settingsGetterContext(args);
+    return !!getters.x1ApiKey && !!state.featureFlags.x1ex;
   },
   chartsFlagEnabled(...args): boolean {
     const { state } = settingsGetterContext(args);
