@@ -1,10 +1,11 @@
 import { FPNumber } from '@sora-substrate/util';
 
-import { KycStatus, VerificationStatus } from '@/types/card';
+import { KycStatus, VerificationStatus, UserInfo, AttemptCounter } from '@/types/card';
 
 import type { Subscription } from 'rxjs';
 
 export type SoraCardState = {
+  kycAttemptCost: string;
   kycStatus: Nullable<KycStatus>;
   verificationStatus: Nullable<VerificationStatus>;
   euroBalance: string;
@@ -13,7 +14,8 @@ export type SoraCardState = {
   xorToDeposit: FPNumber;
   totalXorBalanceUpdates: Nullable<Subscription>;
   authLogin: any;
-  hasFreeAttempts: Nullable<boolean>;
   wantsToPassKycAgain: boolean;
   rejectReason: Nullable<string>;
+  userInfo: UserInfo;
+  attemptCounter: AttemptCounter;
 };
