@@ -1,18 +1,7 @@
 import type { LiquiditySourceTypes } from '@sora-substrate/liquidity-proxy/build/consts';
-import type {
-  QuotePaths,
-  QuotePayload,
-  PrimaryMarketsEnabledAssets,
-  LPRewardsInfo,
-} from '@sora-substrate/liquidity-proxy/build/types';
+import type { LPRewardsInfo, SwapQuote } from '@sora-substrate/liquidity-proxy/build/types';
 import type { CodecString } from '@sora-substrate/util';
 import type { AccountBalance } from '@sora-substrate/util/build/assets/types';
-
-export type DexQuoteData = {
-  pairLiquiditySources: Array<LiquiditySourceTypes>;
-  paths: QuotePaths;
-  payload: Nullable<QuotePayload>;
-};
 
 export type SwapState = {
   tokenFromAddress: Nullable<string>;
@@ -24,10 +13,11 @@ export type SwapState = {
   amountWithoutImpact: CodecString;
   liquidityProviderFee: CodecString;
   isExchangeB: boolean;
-  enabledAssets: PrimaryMarketsEnabledAssets;
   rewards: readonly LPRewardsInfo[];
   route: readonly string[];
+  isAvailable: boolean;
+  liquiditySources: LiquiditySourceTypes[];
+  swapQuote: Nullable<SwapQuote>;
   selectedDexId: number;
-  dexQuoteData: Record<number, DexQuoteData>;
   allowLossPopup: boolean;
 };
