@@ -1,4 +1,5 @@
 import { Operation, isBridgeOperation, isEvmOperation, isSubstrateOperation } from '@sora-substrate/util';
+import { api as soraApi } from '@soramitsu/soraneo-wallet-web';
 import { ethers } from 'ethers';
 
 import { isUnsignedTx as isUnsignedEthTx } from '@/utils/bridge/eth/utils';
@@ -11,41 +12,27 @@ import type { IBridgeTransaction, BridgeHistory } from '@sora-substrate/util';
 import type { EvmHistory } from '@sora-substrate/util/build/bridgeProxy/evm/types';
 import type { SubHistory } from '@sora-substrate/util/build/bridgeProxy/sub/types';
 
-export const waitForEvmTransactionStatus = async (
+const waitForEvmTransactionStatus = async (
   hash: string,
   replaceCallback: (hash: string) => any,
   cancelCallback: (hash: string) => any
 ) => {
-  const ethersInstance = await ethersUtil.getEthersInstance();
   try {
-    // the confirmations value was reduced from 5 to 1
-    // since after the block release, we can be sure that the transaction is completed
-    const confirmations = 1;
-    const timeout = 0;
-    const currentBlock = await ethersInstance.getBlockNumber();
-    const blockOffset = currentBlock - 20;
-    const { data, from, nonce, to, value } = await ethersInstance.getTransaction(hash);
-    await ethersInstance._waitForTransaction(hash, confirmations, timeout, {
-      data,
-      from,
-      nonce,
-      to: to ?? '',
-      value,
-      startBlock: blockOffset,
-    });
+    const ethersInstance = await ethersUtil.getEthersInstance();
+    await ethersInstance.waitForTransaction(hash);
   } catch (error: any) {
-    if (error.code === ethers.errors.TRANSACTION_REPLACED) {
-      if (error.reason === 'repriced' || error.reason === 'replaced') {
-        replaceCallback(error.replacement.hash);
-      } else if (error.reason === 'canceled') {
+    if (ethers.isError(error, 'TRANSACTION_REPLACED')) {
+      if (error.reason === 'cancelled') {
         cancelCallback(error.replacement.hash);
+      } else {
+        replaceCallback(error.replacement.hash);
       }
     }
   }
 };
 
 export const waitForEvmTransactionMined = async (hash?: string, updatedCallback?: (hash: string) => void) => {
-  if (!hash) throw new Error('[Bridge]: evm hash cannot be empty!');
+  if (!hash) throw new Error('[waitForEvmTransactionMined]: hash cannot be empty!');
 
   await waitForEvmTransactionStatus(
     hash,
@@ -54,7 +41,7 @@ export const waitForEvmTransactionMined = async (hash?: string, updatedCallback?
       await waitForEvmTransactionMined(replaceHash, updatedCallback);
     },
     (cancelHash) => {
-      throw new Error(`[Bridge]: The transaction was canceled by the user [${cancelHash}]`);
+      throw new Error(`[waitForEvmTransactionMined]: The transaction was canceled by the user [${cancelHash}]`);
     }
   );
 };
@@ -63,11 +50,13 @@ export const getEvmTransactionRecieptByHash = async (
   transactionHash: string
 ): Promise<{ fee: string; blockHash: string; blockNumber: number; from: string } | null> => {
   try {
-    const { from, effectiveGasPrice, gasUsed, blockNumber, blockHash } = await ethersUtil.getEvmTransactionReceipt(
-      transactionHash
-    );
+    const receipt = await ethersUtil.getEvmTransactionReceipt(transactionHash);
 
-    const fee = ethersUtil.calcEvmFee(effectiveGasPrice.toNumber(), gasUsed.toNumber());
+    if (!receipt) throw new Error(`Transaction receipt "${transactionHash}" not found`);
+
+    const { from, gasPrice, gasUsed, blockNumber, blockHash } = receipt;
+
+    const fee = ethersUtil.calcEvmFee(gasPrice, gasUsed);
 
     return { fee, blockHash, blockNumber, from };
   } catch (error) {
@@ -75,58 +64,18 @@ export const getEvmTransactionRecieptByHash = async (
   }
 };
 
-export const findUserTxIdInBlock = async (
-  api: ApiPromise,
-  blockHash: string,
-  soraHash: string,
-  eventMethod: string,
-  eventSection: string
-): Promise<string | undefined> => {
-  const blockEvents = await getBlockEvents(api, blockHash);
+export const getTransactionEvents = async (blockHash: string, transactionHash: string, api: ApiPromise) => {
+  const extrinsics = await soraApi.system.getExtrinsicsFromBlock(blockHash, api);
+  const extrinsicIndex = extrinsics.findIndex((ext) => ext.hash.toString() === transactionHash);
 
-  const event = blockEvents.find(
-    ({ phase, event }) =>
-      phase.isApplyExtrinsic &&
-      event.section === eventSection &&
-      event.method === eventMethod &&
-      event.data?.[0]?.toString() === soraHash
+  if (extrinsicIndex === -1) throw new Error(`Unable to find extrinsic "${transactionHash}" in block "${blockHash}"`);
+
+  const blockEvents = await soraApi.system.getBlockEvents(blockHash, api);
+  const transactionEvents = blockEvents.filter(
+    ({ phase }) => phase.isApplyExtrinsic && phase.asApplyExtrinsic.toNumber() === extrinsicIndex
   );
 
-  if (!event) return undefined;
-
-  const index = event.phase.asApplyExtrinsic.toNumber();
-  const extrinsics = await getBlockExtrinsics(api, blockHash);
-  const userExtrinsic = extrinsics[index];
-
-  return userExtrinsic.hash.toString();
-};
-
-export const getBlockHash = async (api: ApiPromise, blockNumber: number): Promise<string> => {
-  if (!blockNumber) return '';
-
-  const result = await api.rpc.chain.getBlockHash(blockNumber);
-
-  return result.toString();
-};
-
-export const getBlockExtrinsics = async (api: ApiPromise, blockHash: string) => {
-  const signedBlock = await api.rpc.chain.getBlock(blockHash);
-
-  return signedBlock.block?.extrinsics.toArray() ?? [];
-};
-
-export const getBlockEvents = async (api: ApiPromise, blockHash: string) => {
-  const apiInstanceAtBlock = await api.at(blockHash);
-  const blockEvents = (await apiInstanceAtBlock.query.system.events()).toArray();
-
-  return blockEvents;
-};
-
-export const getBlockTimestamp = async (api: ApiPromise, blockHash: string): Promise<number> => {
-  const apiInstanceAtBlock = await api.at(blockHash);
-  const timestamp = await apiInstanceAtBlock.query.timestamp.now();
-
-  return timestamp.toNumber();
+  return transactionEvents;
 };
 
 export const findEventInBlock = async ({
@@ -140,7 +89,7 @@ export const findEventInBlock = async ({
   section: string;
   method: string;
 }) => {
-  const blockEvents = await getBlockEvents(api, blockId);
+  const blockEvents = await soraApi.system.getBlockEvents(blockId, api);
   const event = blockEvents.find(({ event }) => event.section === section && event.method === method);
 
   if (!event) throw new Error('Event not found');
