@@ -8,7 +8,9 @@ import { ZeroStringValue } from '@/consts';
 import type { SubNetworkApps } from '@/store/web3/types';
 import { subBridgeApi } from '@/utils/bridge/sub/api';
 
-import type { ApiPromise } from '@polkadot/api';
+import type { ApiPromise, ApiRx } from '@polkadot/api';
+import type { SubmittableExtrinsic } from '@polkadot/api-base/types';
+import type { ISubmittableResult } from '@polkadot/types/types';
 import type { CodecString } from '@sora-substrate/util';
 import type { RegisteredAsset } from '@sora-substrate/util/build/assets/types';
 
@@ -16,8 +18,10 @@ export class SubAdapter {
   protected endpoint!: string;
 
   public connection!: Connection;
+  public readonly subNetwork!: SubNetwork;
 
-  constructor() {
+  constructor(subNetwork: SubNetwork) {
+    this.subNetwork = subNetwork;
     this.connection = new Connection({});
   }
 
@@ -26,7 +30,7 @@ export class SubAdapter {
   }
 
   get apiRx() {
-    return this.api.rx;
+    return this.api.rx as ApiRx;
   }
 
   get connected(): boolean {
@@ -49,6 +53,13 @@ export class SubAdapter {
     }
   }
 
+  public getSoraParachainId(): number | undefined {
+    const soraParachain = subBridgeApi.getSoraParachain(this.subNetwork);
+    const soraParachainId = subBridgeApi.parachainIds[soraParachain];
+
+    return soraParachainId;
+  }
+
   public async getBlockNumber(): Promise<number> {
     if (!this.connected) return 0;
 
@@ -58,7 +69,7 @@ export class SubAdapter {
   }
 
   protected async getAccountBalance(accountAddress: string): Promise<CodecString> {
-    if (!this.connected) return ZeroStringValue;
+    if (!(this.connected && accountAddress)) return ZeroStringValue;
 
     const accountInfo = await this.api.query.system.account(accountAddress);
     const accountBalance = formatBalance(accountInfo.data);
@@ -68,26 +79,42 @@ export class SubAdapter {
   }
 
   public async getTokenBalance(accountAddress: string, tokenAddress?: string): Promise<CodecString> {
-    console.info(`[${this.constructor.name}] getTokenBalance method is not implemented`);
-    return ZeroStringValue;
+    throw new Error(`[${this.constructor.name}] "getTokenBalance" method is not implemented`);
   }
 
-  public async transfer(asset: RegisteredAsset, recipient: string, amount: string | number, historyId?: string) {
-    console.info(`[${this.constructor.name}] transfer method is not implemented`);
+  protected getTransferExtrinsic(
+    asset: RegisteredAsset,
+    recipient: string,
+    amount: CodecString
+  ): SubmittableExtrinsic<'promise', ISubmittableResult> {
+    throw new Error(`[${this.constructor.name}] "getTransferExtrinsic" method is not implemented`);
   }
 
+  public async transfer(
+    asset: RegisteredAsset,
+    recipient: string,
+    amount: string | number,
+    historyId?: string
+  ): Promise<void> {
+    throw new Error(`[${this.constructor.name}] "transfer" method is not implemented`);
+  }
+
+  /* [Substrate 5] Runtime call transactionPaymentApi */
   public async getNetworkFee(): Promise<CodecString> {
-    console.info(`[${this.constructor.name}] getNetworkFee method is not implemented`);
-    return ZeroStringValue;
+    const decimals = this.api.registry.chainDecimals[0];
+    const tx = this.getTransferExtrinsic({} as RegisteredAsset, '', ZeroStringValue);
+    const res = await tx.paymentInfo('');
+
+    return new FPNumber(res.partialFee, decimals).toCodecString();
   }
 }
 
-class RococoAdapter extends SubAdapter {
+class KusamaAdapter extends SubAdapter {
   public async getTokenBalance(accountAddress: string, tokenAddress?: string): Promise<CodecString> {
     return await this.getAccountBalance(accountAddress);
   }
 
-  protected getTransferExtrinsic(value: CodecString, recipient: string) {
+  protected getTransferExtrinsic(asset: RegisteredAsset, recipient: string, amount: CodecString) {
     return this.api.tx.xcmPallet.reserveTransferAssets(
       // dest
       {
@@ -95,7 +122,7 @@ class RococoAdapter extends SubAdapter {
           parents: 0,
           interior: {
             X1: {
-              Parachain: subBridgeApi.parachainIds[SubNetwork.RococoSora],
+              Parachain: this.getSoraParachainId(),
             },
           },
         },
@@ -124,7 +151,7 @@ class RococoAdapter extends SubAdapter {
               },
             },
             fun: {
-              Fungible: value,
+              Fungible: amount,
             },
           },
         ],
@@ -134,20 +161,14 @@ class RococoAdapter extends SubAdapter {
     );
   }
 
+  /* Throws error until Substrate 5 migration */
   public async getNetworkFee(): Promise<CodecString> {
-    /* Runtime call transactionPaymentApi not works, not decorated? */
-
-    // try {
-    //   const tx = this.getTransferExtrinsic(ZeroStringValue, '');
-    //   const res = await tx.paymentInfo('');
-
-    //   return new FPNumber(res.partialFee, 12).toCodecString();
-    // } catch (error) {
-    //   console.error(error);
-    //   return ZeroStringValue;
-    // }
-
-    return ZeroStringValue;
+    try {
+      return await super.getNetworkFee();
+    } catch {
+      // Hardcoded value for Kusama - 0.0007 KSM
+      return '700000000';
+    }
   }
 
   public async transfer(asset: RegisteredAsset, recipient: string, amount: string | number, historyId?: string) {
@@ -158,63 +179,11 @@ class RococoAdapter extends SubAdapter {
       symbol: asset.symbol,
       assetAddress: asset.address,
       amount: `${amount}`,
-      externalNetwork: SubNetwork.Rococo,
+      externalNetwork: this.subNetwork,
       externalNetworkType: BridgeNetworkType.Sub,
     };
 
-    const extrinsic = this.getTransferExtrinsic(value, recipient);
-
-    await subBridgeApi.submitApiExtrinsic(this.api, extrinsic, subBridgeApi.account.pair, historyItem);
-  }
-}
-
-/** Not used & tested yet */
-class KusamaKaruraAdapter extends SubAdapter {
-  // [TODO] fetch balance by symbol
-  public async getTokenBalance(accountAddress: string, tokenAddress?: string): Promise<CodecString> {
-    return await this.getAccountBalance(accountAddress);
-  }
-
-  public async transfer(asset: RegisteredAsset, recipient: string, amount: string | number, historyId?: string) {
-    const value = new FPNumber(amount, asset.externalDecimals).toCodecString();
-
-    const historyItem = subBridgeApi.getHistory(historyId as string) || {
-      type: Operation.SubstrateIncoming,
-      symbol: asset.symbol,
-      assetAddress: asset.address,
-      amount: `${amount}`,
-      externalNetwork: SubNetwork.KusamaKarura,
-      externalNetworkType: BridgeNetworkType.Sub,
-    };
-
-    const extrinsic = this.api.tx.xTokens.transfer(
-      // currencyId: AcalaPrimitivesCurrencyCurrencyId
-      {
-        Token: asset.symbol,
-      },
-      // amount: u128 (Balance)
-      value,
-      // dest: XcmVersionedMultiLocation
-      {
-        V1: {
-          parents: 1,
-          interior: {
-            X2: [
-              {
-                Parachain: subBridgeApi.parachainIds[SubNetwork.KusamaSora],
-              },
-              {
-                AccountId32: {
-                  id: this.api.createType('AccountId32', recipient).toHex(),
-                },
-              },
-            ],
-          },
-        },
-      },
-      // destWeightLimit
-      'Unlimited'
-    );
+    const extrinsic = this.getTransferExtrinsic(asset, recipient, value);
 
     await subBridgeApi.submitApiExtrinsic(this.api, extrinsic, subBridgeApi.account.pair, historyItem);
   }
@@ -222,11 +191,10 @@ class KusamaKaruraAdapter extends SubAdapter {
 
 class SubConnector {
   public readonly adapters = {
-    [SubNetwork.Rococo]: () => new RococoAdapter(),
-    [SubNetwork.RococoSora]: () => new SubAdapter(),
-    /** Not used yet */
-    // [SubNetwork.KusamaKarura]: new KusamaKaruraAdapter(),
-    // [SubNetwork.KusamaSora]: new SubAdapter(),
+    [SubNetwork.Rococo]: () => new KusamaAdapter(SubNetwork.Rococo),
+    [SubNetwork.Kusama]: () => new KusamaAdapter(SubNetwork.Kusama),
+    [SubNetwork.RococoSora]: () => new SubAdapter(SubNetwork.RococoSora),
+    [SubNetwork.KusamaSora]: () => new SubAdapter(SubNetwork.KusamaSora),
   };
 
   public endpoints: SubNetworkApps = {};
