@@ -1,4 +1,5 @@
 import { FPNumber, CodecString } from '@sora-substrate/util';
+import { isNativeAsset } from '@sora-substrate/util/build/assets';
 import { XOR } from '@sora-substrate/util/build/assets/consts';
 import { api, WALLET_CONSTS } from '@soramitsu/soraneo-wallet-web';
 import debounce from 'lodash/debounce';
@@ -15,7 +16,9 @@ import type { Asset, AccountAsset, RegisteredAccountAsset } from '@sora-substrat
 import type { AccountLiquidity } from '@sora-substrate/util/build/poolXyk/types';
 import type { Route } from 'vue-router';
 
-type AssetWithBalance = AccountAsset | AccountLiquidity | RegisteredAccountAsset;
+type AssetWithBalance = AccountAsset | RegisteredAccountAsset;
+
+type PoolAssets<T extends Asset> = { baseAsset: T; poolAsset: T };
 
 export async function waitForSoraNetworkFromEnv(): Promise<WALLET_CONSTS.SoraNetwork> {
   return new Promise<WALLET_CONSTS.SoraNetwork>((resolve) => {
@@ -46,28 +49,19 @@ export const isXorAccountAsset = (asset: Asset | AssetWithBalance): boolean => {
   return asset ? asset.address === XOR.address : false;
 };
 
-export const isNativeEvmTokenAddress = (address: string): boolean => {
-  const numberString = address.replace(/^0x/, '');
-  const number = parseInt(numberString, 16);
-
-  return number === 0;
-};
-
 export const isMaxButtonAvailable = (
-  areAssetsSelected: boolean,
   asset: AssetWithBalance,
   amount: string | number,
   fee: CodecString,
   xorAsset: AccountAsset | RegisteredAccountAsset,
-  parseAsLiquidity = false,
   isXorOutputSwap = false
 ): boolean => {
-  if (!asset || !areAssetsSelected || !xorAsset || asZeroValue(getAssetBalance(asset, { parseAsLiquidity }))) {
+  if (!asset || !xorAsset || asZeroValue(getAssetBalance(asset))) {
     return false;
   }
 
   const fpAmount = new FPNumber(amount, asset.decimals);
-  const fpMaxBalance = getMaxBalance(asset, fee, false, parseAsLiquidity);
+  const fpMaxBalance = getMaxBalance(asset, fee);
 
   return !FPNumber.eq(fpMaxBalance, fpAmount) && !hasInsufficientXorForFee(xorAsset, fee, isXorOutputSwap);
 };
@@ -75,24 +69,21 @@ export const isMaxButtonAvailable = (
 const getMaxBalance = (
   asset: AssetWithBalance,
   fee: CodecString,
-  isExternalBalance = false,
-  parseAsLiquidity = false,
-  isBondedBalance = false
+  { isExternalBalance = false, isExternalNative = false, isBondedBalance = false } = {}
 ): FPNumber => {
-  const balance = getAssetBalance(asset, { internal: !isExternalBalance, parseAsLiquidity, isBondedBalance });
-  const decimals: number = asset[isExternalBalance ? 'externalDecimals' : 'decimals'];
+  const balance = getAssetBalance(asset, { internal: !isExternalBalance, isBondedBalance });
+  const decimals = getAssetDecimals(asset, { internal: !isExternalBalance }) as number;
 
   if (asZeroValue(balance)) return FPNumber.ZERO;
 
   let fpResult = FPNumber.fromCodecValue(balance, decimals);
 
   if (
+    !isBondedBalance &&
     !asZeroValue(fee) &&
-    ((!isExternalBalance && isXorAccountAsset(asset)) ||
-      (isExternalBalance && isNativeEvmTokenAddress((asset as RegisteredAccountAsset).externalAddress))) &&
-    !isBondedBalance
+    ((!isExternalBalance && isXorAccountAsset(asset)) || (isExternalBalance && isExternalNative))
   ) {
-    const fpFee = FPNumber.fromCodecValue(fee);
+    const fpFee = FPNumber.fromCodecValue(fee, decimals);
     fpResult = fpResult.sub(fpFee);
   }
 
@@ -102,10 +93,9 @@ const getMaxBalance = (
 export const getMaxValue = (
   asset: AccountAsset | RegisteredAccountAsset,
   fee: CodecString,
-  isExternalBalance = false,
-  isBondedBalance = false
+  { isExternalBalance = false, isExternalNative = false, isBondedBalance = false } = {}
 ): string => {
-  return getMaxBalance(asset, fee, isExternalBalance, false, isBondedBalance).toString();
+  return getMaxBalance(asset, fee, { isExternalBalance, isExternalNative, isBondedBalance }).toString();
 };
 
 export const getDeltaPercent = (desiredPrice: FPNumber, currentPrice: FPNumber): FPNumber => {
@@ -121,7 +111,7 @@ export const hasInsufficientBalance = (
   isBondedBalance = false
 ): boolean => {
   const fpAmount = new FPNumber(amount, asset.decimals);
-  const fpMaxBalance = getMaxBalance(asset, fee, isExternalBalance, false, isBondedBalance);
+  const fpMaxBalance = getMaxBalance(asset, fee, { isExternalBalance, isBondedBalance });
 
   return FPNumber.lt(fpMaxBalance, fpAmount);
 };
@@ -142,7 +132,7 @@ export const hasInsufficientXorForFee = (
   return FPNumber.lt(fpBalance, fpFee) && !isXorOutputSwap;
 };
 
-export const hasInsufficientEvmNativeTokenForFee = (nativeBalance: CodecString, fee: CodecString): boolean => {
+export const hasInsufficientNativeTokenForFee = (nativeBalance: CodecString, fee: CodecString): boolean => {
   if (!fee) return false;
 
   const fpBalance = FPNumber.fromCodecValue(nativeBalance);
@@ -165,7 +155,7 @@ export const asZeroValue = (value: any): boolean => {
 
 export const getAssetBalance = (
   asset: Nullable<AssetWithBalance>,
-  { internal = true, parseAsLiquidity = false, isBondedBalance = false } = {}
+  { internal = true, isBondedBalance = false } = {}
 ) => {
   if (!asset) return ZeroStringValue;
 
@@ -173,15 +163,15 @@ export const getAssetBalance = (
     return (asset as RegisteredAccountAsset)?.externalBalance;
   }
 
-  if (parseAsLiquidity) {
-    return (asset as AccountLiquidity)?.balance;
-  }
-
   if (isBondedBalance) {
     return (asset as AccountAsset)?.balance?.bonded;
   }
 
   return (asset as AccountAsset)?.balance?.transferable;
+};
+
+export const getLiquidityBalance = (liquidity: Nullable<AccountLiquidity>): CodecString | undefined => {
+  return liquidity?.balance;
 };
 
 export const getAssetDecimals = (asset: any, { internal = true } = {}): number | undefined => {
@@ -192,17 +182,11 @@ export const getAssetDecimals = (asset: any, { internal = true } = {}): number |
 
 export const formatAssetBalance = (
   asset: any,
-  {
-    internal = true,
-    parseAsLiquidity = false,
-    formattedZero = '',
-    showZeroBalance = true,
-    isBondedBalance = false,
-  } = {}
+  { internal = true, formattedZero = '', showZeroBalance = true, isBondedBalance = false } = {}
 ): string => {
   if (!asset) return formattedZero;
 
-  const balance = getAssetBalance(asset, { internal, parseAsLiquidity, isBondedBalance });
+  const balance = getAssetBalance(asset, { internal, isBondedBalance });
 
   if (!balance || (!showZeroBalance && asZeroValue(balance))) return formattedZero;
 
@@ -308,4 +292,33 @@ export const formatDecimalPlaces = (value: FPNumber | number, asPercent = false)
   const postfix = asPercent ? '%' : '';
 
   return `${formatted}${postfix}`;
+};
+
+const sortAssetsByProp = <T extends Asset>(a: T, b: T, prop: 'address' | 'symbol' | 'name') => {
+  if (a[prop] < b[prop]) return -1;
+  if (a[prop] > b[prop]) return 1;
+  return 0;
+};
+
+export const sortAssets = <T extends Asset>(a: T, b: T) => {
+  const isNativeA = isNativeAsset(a);
+  const isNativeB = isNativeAsset(b);
+  // sort native assets by address
+  if (isNativeA && isNativeB) {
+    return sortAssetsByProp(a, b, 'address');
+  }
+  if (isNativeA && !isNativeB) {
+    return -1;
+  }
+  if (!isNativeA && isNativeB) {
+    return 1;
+  }
+  // sort non native assets by symbol
+  return sortAssetsByProp(a, b, 'symbol');
+};
+
+export const sortPools = <T extends Asset>(a: PoolAssets<T>, b: PoolAssets<T>) => {
+  const byBaseAsset = sortAssets(a.baseAsset, b.baseAsset);
+
+  return byBaseAsset === 0 ? sortAssets(a.poolAsset, b.poolAsset) : byBaseAsset;
 };
