@@ -172,7 +172,7 @@ import { demeterLazyComponent } from '@/modules/demeterFarming/router';
 import type { DemeterPoolDerivedData } from '@/modules/demeterFarming/types';
 import { lazyComponent } from '@/router';
 import type { AmountWithSuffix } from '@/types/formats';
-import { formatAmountWithSuffix, formatDecimalPlaces } from '@/utils';
+import { formatAmountWithSuffix, formatDecimalPlaces, sortPools } from '@/utils';
 
 import type { Asset } from '@sora-substrate/util/build/assets/types';
 import type { DemeterPool } from '@sora-substrate/util/build/demeterFarming/types';
@@ -244,7 +244,7 @@ export default class ExploreDemeter extends Mixins(TranslationMixin, DemeterBase
   }
 
   get items(): TableItem[] {
-    return this.list.map((pool) => {
+    const items = this.list.map((pool) => {
       const baseAsset = this.demeterAssetsData[pool.baseAsset] as Asset;
       const poolAsset = this.demeterAssetsData[pool.poolAsset] as Asset;
       const rewardAsset = this.demeterAssetsData[pool.rewardAsset] as Asset;
@@ -325,12 +325,19 @@ export default class ExploreDemeter extends Mixins(TranslationMixin, DemeterBase
         liquidity,
       };
     });
+
+    const defaultSorted = [...items].sort((a, b) =>
+      sortPools(
+        { baseAsset: a.poolAsset, poolAsset: a.rewardAsset },
+        { baseAsset: b.poolAsset, poolAsset: b.rewardAsset }
+      )
+    );
+
+    return defaultSorted;
   }
 
   get preparedItems(): TableItem[] {
-    if (!this.isAccountItemsOnly) return this.items;
-
-    return this.items.filter((item) => item.isAccountItem);
+    return this.isAccountItemsOnly ? this.items.filter((item) => item.isAccountItem) : this.items;
   }
 
   // ExplorePageMixin method implementation
