@@ -1,12 +1,6 @@
 <template>
   <div>
-    <div class="switcher">
-      <s-switch v-model="isSynthsOnly" />
-      <span>
-        {{ t('explore.showOnly') }}
-        <external-link default-class="p3" :title="t('explore.synthetics')" :href="SYNTHS_LINK" />
-      </span>
-    </div>
+    <synthetic-switcher v-model="isSynthsOnly" />
     <s-table
       ref="table"
       v-loading="loadingState"
@@ -184,7 +178,7 @@ import TranslationMixin from '@/components/mixins/TranslationMixin';
 import { Components } from '@/consts';
 import { lazyComponent } from '@/router';
 import { getter } from '@/store/decorators';
-import { calcPriceChange, formatAmountWithSuffix } from '@/utils';
+import { calcPriceChange, formatAmountWithSuffix, sortAssets } from '@/utils';
 import { syntheticAssetRegexp } from '@/utils/regexp';
 import storage from '@/utils/storage';
 
@@ -287,7 +281,7 @@ const parse = (item: AssetData): Record<string, TokenData> => {
 
 @Component({
   components: {
-    ExternalLink: lazyComponent(Components.ExternalLink),
+    SyntheticSwitcher: lazyComponent(Components.SyntheticSwitcher),
     PriceChange: lazyComponent(Components.PriceChange),
     SortButton: lazyComponent(Components.SortButton),
     TokenAddress: components.TokenAddress,
@@ -298,10 +292,8 @@ const parse = (item: AssetData): Record<string, TokenData> => {
 })
 export default class Tokens extends Mixins(ExplorePageMixin, TranslationMixin) {
   private readonly DAY = 60 * 60 * 24;
-  readonly SYNTHS_LINK =
-    'https://medium.com/polkaswap/unveiling-synthetic-assets-a-game-changer-in-the-financial-landscape-1720e5858422';
 
-  @getter.assets.whitelistAssets private items!: Array<Asset>;
+  @getter.assets.whitelistAssets private whitelistAssets!: Array<Asset>;
 
   private isSynths = storage.get(storageKey as any) ? JSON.parse(storage.get(storageKey as any)) : false;
 
@@ -323,45 +315,50 @@ export default class Tokens extends Mixins(ExplorePageMixin, TranslationMixin) {
     return Object.keys(this.tokensData).length !== 0;
   }
 
+  get items(): TableItem[] {
+    const items = Object.entries(this.tokensData).reduce<TableItem[]>((buffer, [address, tokenData]) => {
+      const asset = this.getAsset(address);
+
+      if (!asset) return buffer;
+
+      const fpPrice = FPNumber.fromCodecValue(this.getAssetFiatPrice(asset) ?? 0);
+      const fpPriceDay = tokenData?.startPriceDay ?? FPNumber.ZERO;
+      const fpPriceWeek = tokenData?.startPriceWeek ?? FPNumber.ZERO;
+      const fpVolumeDay = tokenData?.volumeDay ?? FPNumber.ZERO;
+      const fpVolumeWeek = tokenData?.volumeWeek ?? FPNumber.ZERO;
+      const fpPriceChangeDay = calcPriceChange(fpPrice, fpPriceDay);
+      const fpPriceChangeWeek = calcPriceChange(fpPrice, fpPriceWeek);
+
+      const reserves = tokenData?.reserves ?? FPNumber.ZERO;
+      const tvl = reserves.mul(fpPrice);
+      const velocity = tvl.isZero() ? FPNumber.ZERO : fpVolumeWeek.div(tvl);
+
+      buffer.push({
+        ...asset,
+        price: fpPrice.toNumber(),
+        priceFormatted: new FPNumber(fpPrice.toFixed(7)).toLocaleString(),
+        priceChangeDay: fpPriceChangeDay.toNumber(),
+        priceChangeDayFP: fpPriceChangeDay,
+        priceChangeWeek: fpPriceChangeWeek.toNumber(),
+        priceChangeWeekFP: fpPriceChangeWeek,
+        volumeDay: fpVolumeDay.toNumber(),
+        volumeDayFormatted: formatAmountWithSuffix(fpVolumeDay),
+        tvl: tvl.toNumber(),
+        tvlFormatted: formatAmountWithSuffix(tvl),
+        velocity: velocity.toNumber(),
+        velocityFormatted: String(velocity.toNumber(2)),
+      });
+
+      return buffer;
+    }, []);
+
+    const defaultSorted = [...items].sort((a, b) => sortAssets(a, b));
+
+    return defaultSorted;
+  }
+
   get preparedItems(): TableItem[] {
-    return this.items // TODO: [PW-1166] map fn is used here cuz whitelistAssets has default sorting
-      .map<[string, TokenData]>(({ address }) => [address, this.tokensData[address]])
-      .reduce<TableItem[]>((buffer, [address, tokenData]) => {
-        const asset = this.getAsset(address);
-
-        if (!asset) return buffer;
-        if (this.isSynthsOnly && !this.isSynthetic(asset.address)) return buffer;
-
-        const fpPrice = FPNumber.fromCodecValue(this.getAssetFiatPrice(asset) ?? 0);
-        const fpPriceDay = tokenData?.startPriceDay ?? FPNumber.ZERO;
-        const fpPriceWeek = tokenData?.startPriceWeek ?? FPNumber.ZERO;
-        const fpVolumeDay = tokenData?.volumeDay ?? FPNumber.ZERO;
-        const fpVolumeWeek = tokenData?.volumeWeek ?? FPNumber.ZERO;
-        const fpPriceChangeDay = calcPriceChange(fpPrice, fpPriceDay);
-        const fpPriceChangeWeek = calcPriceChange(fpPrice, fpPriceWeek);
-
-        const reserves = tokenData?.reserves ?? FPNumber.ZERO;
-        const tvl = reserves.mul(fpPrice);
-        const velocity = tvl.isZero() ? FPNumber.ZERO : fpVolumeWeek.div(tvl);
-
-        buffer.push({
-          ...asset,
-          price: fpPrice.toNumber(),
-          priceFormatted: new FPNumber(fpPrice.toFixed(7)).toLocaleString(),
-          priceChangeDay: fpPriceChangeDay.toNumber(),
-          priceChangeDayFP: fpPriceChangeDay,
-          priceChangeWeek: fpPriceChangeWeek.toNumber(),
-          priceChangeWeekFP: fpPriceChangeWeek,
-          volumeDay: fpVolumeDay.toNumber(),
-          volumeDayFormatted: formatAmountWithSuffix(fpVolumeDay),
-          tvl: tvl.toNumber(),
-          tvlFormatted: formatAmountWithSuffix(tvl),
-          velocity: velocity.toNumber(),
-          velocityFormatted: String(velocity.toNumber(2)),
-        });
-
-        return buffer;
-      }, []);
+    return this.isSynthsOnly ? this.items.filter((item) => this.isSynthetic(item.address)) : this.items;
   }
 
   isSynthetic(address: string): boolean {
@@ -381,7 +378,7 @@ export default class Tokens extends Mixins(ExplorePageMixin, TranslationMixin) {
     const now = Math.floor(Date.now() / (5 * 60_000)) * (5 * 60); // rounded to latest 5min snapshot (unix)
     const dayTimestamp = now - this.DAY; // latest day snapshot (unix)
     const weekTimestamp = now - this.DAY * 7; // latest week snapshot (unix)
-    const ids = this.items.map((item) => item.address); // only whitelisted assets
+    const ids = this.whitelistAssets.map((item) => item.address); // only whitelisted assets
 
     const variables = { ids, dayTimestamp, weekTimestamp };
     const items = await SubqueryExplorerService.fetchAllEntities(AssetsQuery, variables, parse);
@@ -395,15 +392,4 @@ export default class Tokens extends Mixins(ExplorePageMixin, TranslationMixin) {
 
 <style lang="scss">
 @include explore-table;
-</style>
-
-<style lang="scss" scoped>
-.switcher {
-  display: flex;
-  align-items: center;
-
-  & > span {
-    margin-left: $inner-spacing-small;
-  }
-}
 </style>
